refactored experiments
