+++ few_shot/experiments/fashion/__init__.py
@@ -100,8 +100,7 @@ def evaluate_fashion_few_shot(train_df,
                             steps_per_epoch=eps_per_epoch,
                             shuffle=False,
                             callbacks=callbacks,
-                            verbose=1
-)
+                            verbose=1)
 
         print(history.history)
         latest_weights = model.get_weights()
